+++ pennylane/tape/tape.py
@@ -1412,27 +1412,6 @@ class QuantumTape(AnnotatedQueue):
     _execute = execute_device
 
 
-class FunctionTape(QuantumTape):
-
-    def __init__(self, param_names, *args, **kwargs):
-        self.param_names = param_names
-        self.values = None
-        self.vars = {param_names[i]:(lambda: self.values[i]) for i in range(len(param_names))}
-        super().__init__(*args, **kwargs)
-
-    def __call__(self, *args):
-        self.values = args
-
-class IfTape(QuantumTape):
-
-    def __init__(self, expr, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.expr = expr
-
-class WhileTape(QuantumTape):
-    def __init__(self, expr, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.expr = expr
 
 def remove_control_flow_transform(tape: QuantumTape):
     with QuantumTape() as new_tape:
@@ -1441,7 +1420,7 @@ def remove_control_flow_transform(tape: QuantumTape):
                 if op.expr():
                     for if_op in op.queue:
                         qml.apply(if_op)
-            if isinstance(op, WhileTape):
+            elif isinstance(op, WhileTape):
                 while op.expr():
                     for while_op in op.queue:
                         qml.apply(while_op)
